+++ cmd/projectList.go
@@ -15,20 +15,74 @@
 package cmd
 
 import (
-	"fmt"
+	"io"
+	"os"
+	"strings"
 
+	"github.com/lucassabreu/clockify-cli/api"
+	"github.com/lucassabreu/clockify-cli/api/dto"
+	"github.com/lucassabreu/clockify-cli/reports"
 	"github.com/spf13/cobra"
+	"github.com/spf13/viper"
 )
 
 // projectListCmd represents the projectList command
 var projectListCmd = &cobra.Command{
 	Use:   "list",
 	Short: "List projects on Clockify and project links",
-	Run: func(cmd *cobra.Command, args []string) {
-		fmt.Println("projectList called")
-	},
+	Run: withClockifyClient(func(cmd *cobra.Command, args []string, c *api.Client) {
+		format, _ := cmd.Flags().GetString("format")
+		quiet, _ := cmd.Flags().GetBool("quiet")
+
+		projects, err := c.GetProjects(api.GetProjectsParam{
+			Workspace: viper.GetString("workspace"),
+		})
+
+		if err != nil {
+			printError(err)
+			return
+		}
+
+		name, _ := cmd.Flags().GetString("name")
+		projects = filterProjects(name, projects)
+
+		var reportFn func([]dto.Project, io.Writer) error
+
+		reportFn = reports.ProjectPrint
+		if format != "" {
+			reportFn = reports.ProjectPrintWithTemplate(format)
+		}
+
+		if quiet {
+			reportFn = reports.ProjectPrintQuietly
+		}
+
+		if err = reportFn(projects, os.Stdout); err != nil {
+			printError(err)
+		}
+	}),
+}
+
+func filterProjects(name string, projects []dto.Project) []dto.Project {
+	if name == "" {
+		return projects
+	}
+
+	ts := make([]dto.Project, 0)
+
+	for _, t := range projects {
+		if strings.Contains(strings.ToLower(t.Name), strings.ToLower(name)) {
+			ts = append(ts, t)
+		}
+	}
+
+	return ts
 }
 
 func init() {
 	projectCmd.AddCommand(projectListCmd)
+
+	projectListCmd.Flags().StringP("name", "n", "", "will be used to filter the tag by name")
+	projectListCmd.Flags().StringP("format", "f", "", "golang text/template format to be applyed on each Project")
+	projectListCmd.Flags().BoolP("quiet", "q", false, "only display ids")
 }